Were refactor __init__.py with blue print. Were submitted separate dir for admin and news files.
