+++ learn_web/create_admin.py
@@ -2,7 +2,7 @@ from getpass import getpass
 import sys
 
 from webapp import create_app
-from webapp.model import User, db
+from webapp.db import User, db
 
 app = create_app()
