+++ learn_web/webapp/admin/views.py
@@ -1,0 +1,13 @@
+from flask_login import current_user, login_required
+from flask import Blueprint
+
+blueprint = Blueprint('admin', __name__, url_prefix='/admin')
+
+
+@blueprint.route('/')
+@login_required
+def admin_index():
+    if current_user.is_admin:
+        return 'Hi, Admin!'
+    else:
+        return 'You are not Admin!'
