+++ learn_web/webapp/news/models.py
@@ -1,6 +1,4 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
+from learn_python.learn_web.webapp.db import db
 
 
 class News(db.Model):
@@ -8,7 +6,7 @@ class News(db.Model):
     title = db.Column(db.String, nullable=False)
     url = db.Column(db.String, unique=True, nullable=False)
     published = db.Column(db.DateTime, nullable=False)
-    text = db.Column(db.Text,nullable=True)
+    text = db.Column(db.Text, nullable=True)
 
     def __repr__(self):
         return '<News {} {}>'.format(self.title, self.url)
