+++ learn_web/webapp/user/models.py
@@ -1,6 +1,6 @@
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
-from webapp.model import db
+from learn_python.learn_web.webapp.db import db
 
 
 class User(db.Model, UserMixin):
