+++ learn_web/webapp/user/views.py
@@ -24,7 +24,7 @@ def process_login():
         if user and user.check_password(form.password.data):
             login_user(user, remember=form.remember_me.data)
             flash('Вы успешно вошли на сайт')
-            return redirect(url_for('index'))
+            return redirect(url_for('news.index'))
     flash('Неправильное имя или пароль')
     return redirect(url_for('user.login'))
 
@@ -33,4 +33,4 @@ def process_login():
 def logout():
     logout_user()
     flash('Пока!')
-    return redirect(url_for('index'))
+    return redirect(url_for('news.index'))
